fixed item-->quest and added a check before deletion

+++ src/main/java/rpgcharacters/commands/CommandDeleteQuest.java
@@ -8,23 +8,38 @@ import java.sql.Statement;
 import com.beust.jcommander.Parameter;
 import com.beust.jcommander.Parameters;
 
-@Parameters(commandDescription = "Create a item")
+@Parameters(commandDescription = "Delete quest")
 public class CommandDeleteQuest implements Command {
     // pass as '... --name "example name" ...'
     @Parameter(names = "--name",
-               description = "Name of the item being created",
+               description = "Name of the quest to delete",
                required = true)
     private String name;
 
     @Override
     public void run(Connection conn) {
         try {
-            String query = "DELETE FROM quest " +
-                           "WHERE name='" + name + "';";
+          String checkQuery = "SELECT * FROM quest" +
+                              "WHERE name='" + name + "';";
 
-            Statement stmt = conn.createStatement();
-            stmt.execute(query);
+          Statement checkStmt = conn.createStatement();
+          ResultSet results = checkStmt.executeQuery(checkQuery);
+
+          results.last();
+          int total = results.getRow();
+
+          if (total == 0) {
+            System.out.println("Quest " + name + " does not exist.");
+          }
+          else {
+            String deleteQuery = "DELETE FROM quest " +
+                                 "WHERE name='" + name + "';";
+
+            Statement deleteStmt = conn.createStatement();
+            deleteStmt.execute(deleteQuery);
+
             System.out.println("Quest " + name + " has been deleted.");
+          }
       } catch (SQLException e) {
           e.printStackTrace();
       }
